Update repository docs with project intentions.
This is a derivative work of the Casper theme by Ghost. Doc updates should make it clear that most of the code in this project is the original work of the Ghost team, and not mine or any future contributor's work.
Retains original LICENSE as required.
Updates metadata and instructions to point at upstream fork where appropriate.
I'm not totally sure of what I'm doing but these changes seem sensible...